@@ -1,4 +1,4 @@
-using System;
+﻿using System;
 
 namespace ca1045
 {
@@ -26,11 +26,12 @@ namespace ca1045
         public static bool ReplyInformation(TypeOfFeedback input,
            out string reply, ref Actions action)
         {
-            bool returnReply = false;
-            string replyText = "Your feedback has been forwarded " +
-                               "to the product manager.";
+            string replyText = """
+                Your feedback has been forwarded to the product manager.
+                """;
 
-            reply = String.Empty;
+            reply = string.Empty;
+            bool returnReply;
             switch (input)
             {
                 case TypeOfFeedback.Complaint:
@@ -57,66 +58,33 @@ namespace ca1045
     // Redesigned version does not use out or ref parameters;
     // instead, it returns this container type.
 
-    public class ReplyData
+    public record class ReplyData(string Reply, Actions Action, bool ReturnReply = false)
     {
-        string reply;
-        Actions action;
-        bool returnReply;
-
-        // Constructors.
-        public ReplyData()
-        {
-            this.reply = String.Empty;
-            this.action = Actions.Discard;
-            this.returnReply = false;
-        }
-
-        public ReplyData(Actions action, string reply, bool returnReply)
-        {
-            this.reply = reply;
-            this.action = action;
-            this.returnReply = returnReply;
-        }
-
-        // Properties.
-        public string Reply { get { return reply; } }
-        public Actions Action { get { return action; } }
-
         public override string ToString()
         {
-            return String.Format("Reply: {0} Action: {1} return? {2}",
-               reply, action.ToString(), returnReply.ToString());
+            return string.Format("Reply: {0} Action: {1} return? {2}",
+               Reply, Action.ToString(), ReturnReply.ToString());
         }
     }
 
     public class RedesignedRefAndOut
     {
-        public static ReplyData ReplyInformation(TypeOfFeedback input)
+        public static ReplyData? ReplyInformation(TypeOfFeedback input)
         {
-            ReplyData answer;
             string replyText = "Your feedback has been forwarded " +
                "to the product manager.";
-
-            switch (input)
+            ReplyData? answer = input switch
             {
-                case TypeOfFeedback.Complaint:
-                case TypeOfFeedback.Praise:
-                    answer = new ReplyData(
-                       Actions.ForwardToManagement,
-                       "Thank you. " + replyText,
-                       true);
-                    break;
-                case TypeOfFeedback.Suggestion:
-                    answer = new ReplyData(
-                       Actions.ForwardToDeveloper,
-                       replyText,
-                       true);
-                    break;
-                case TypeOfFeedback.Incomprehensible:
-                default:
-                    answer = new ReplyData();
-                    break;
-            }
+                TypeOfFeedback.Complaint or TypeOfFeedback.Praise => new ReplyData(
+                                       "Thank you. " + replyText,
+                                       Actions.ForwardToManagement,
+                                       true),
+                TypeOfFeedback.Suggestion => new ReplyData(
+                                       replyText,
+                                       Actions.ForwardToDeveloper,
+                                       true),
+                _ => null,
+            };
             return answer;
         }
     }
@@ -133,13 +101,13 @@ namespace ca1045
             string[] reply = new string[5];
 
             // You must initialize a ref parameter.
-            Actions[] action = {Actions.Unknown,Actions.Unknown,
+            Actions[] action = [Actions.Unknown,Actions.Unknown,
                              Actions.Unknown,Actions.Unknown,
-                             Actions.Unknown,Actions.Unknown};
+                             Actions.Unknown,Actions.Unknown];
             bool[] disposition = new bool[5];
             int i = 0;
 
-            foreach (TypeOfFeedback t in Enum.GetValues(typeof(TypeOfFeedback)))
+            foreach (TypeOfFeedback t in Enum.GetValues<TypeOfFeedback>())
             {
                 // The call to the library.
                 disposition[i] = BadRefAndOut.ReplyInformation(
@@ -153,7 +121,7 @@ namespace ca1045
         {
             ReplyData[] answer = new ReplyData[5];
             int i = 0;
-            foreach (TypeOfFeedback t in Enum.GetValues(typeof(TypeOfFeedback)))
+            foreach (TypeOfFeedback t in Enum.GetValues<TypeOfFeedback>())
             {
                 // The call to the library.
                 answer[i] = RedesignedRefAndOut.ReplyInformation(t);
