@@ -1,5 +1,5 @@
+﻿using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
-using System.Runtime.CompilerServices;
 
 [assembly: DisableRuntimeMarshalling]
 
@@ -7,5 +7,5 @@ class C
 {
     // Violates rule CA1420.
     [DllImport("NativeLibrary", SetLastError = true)]
-    public static extern void MyMethod ();
+    public static extern void MyMethod();
 }
