@@ -1,25 +1,24 @@
-﻿Imports System
-Imports System.Data.SqlClient
+﻿Imports System.IO
 
 Namespace ca1816
 
     '<snippet1>
-    Public Class DatabaseConnector
+    Public Class MyStreamClass
         Implements IDisposable
 
-        Private _Connection As New SqlConnection
+        Private _stream As New MemoryStream
 
         Public Sub Dispose() Implements IDisposable.Dispose
             Dispose(True)
-            ' Violates rules
+            ' Violates rule.
             GC.SuppressFinalize(True)
         End Sub
 
         Protected Overridable Sub Dispose(ByVal disposing As Boolean)
             If disposing Then
-                If _Connection IsNot Nothing Then
-                    _Connection.Dispose()
-                    _Connection = Nothing
+                If _stream IsNot Nothing Then
+                    _stream.Dispose()
+                    _stream = Nothing
                 End If
             End If
         End Sub
@@ -31,10 +30,10 @@ End Namespace
 
 Namespace ca1816_2
     '<snippet2>
-    Public Class DatabaseConnector
+    Public Class MyStreamClass
         Implements IDisposable
 
-        Private _Connection As New SqlConnection
+        Private _stream As New MemoryStream
 
         Public Sub Dispose() Implements IDisposable.Dispose
             Dispose(True)
@@ -43,9 +42,9 @@ Namespace ca1816_2
 
         Protected Overridable Sub Dispose(ByVal disposing As Boolean)
             If disposing Then
-                If _Connection IsNot Nothing Then
-                    _Connection.Dispose()
-                    _Connection = Nothing
+                If _stream IsNot Nothing Then
+                    _stream.Dispose()
+                    _stream = Nothing
                 End If
             End If
         End Sub
